I don't remember
I believe these are changes made during the previous class on Tuesday of the second to last week in November.

I remember that it was about player input, so this code is all related to movement and getting it to work. But last I remembered, the code stopped Unity from starting up the game, so it might be bad or unupdated code.

If everything is right with it though, this is just code that's indirectly copied from the lesson to get the character to move around, I'll change it later, maybe try to use RigidBody2D to allow for player input instead.