Large variety of changes regarding prefabs
I haven't updated this repository in a little while, I'm pretty sure it's all notes and content from class, since I haven't turned in any of the medal tasks. Basically, I made prefabs for the character, made the coin/gem, and turned that into a prefab too. Added a second scene to do testing in, added a coin spawner that can spawn the coins, later they will be able to spawn after a trigger is activated. I currently have an unfinished trigger added, it doesn't do anything, but it functions as I would want it to, where it has a hitbox, but isn't visible. And on the day of writing this, I was taught how to add coins with different points, which also extended to coins that can return a value to the player, increasing the score. This was extended into me adding a power up that increases the speed variable, which is the last thing I did before writing this.